+++ scripts/content.js
@@ -1,12 +1,3 @@
 "use strict";
 
-function normalizeRentList() {
-    const normalizer = new RentItemsNormalizer()
-    normalizer.normalize()
-}
-
-function bootstrap() {
-    normalizeRentList()
-}
-
 bootstrap()